fix negative offsets

+++ x86/x86asm/decode.go
@@ -537,7 +537,7 @@ Decode:
 						if pos+2 > len(src) {
 							return truncated(src, mode)
 						}
-						mem.Disp = int64(binary.LittleEndian.Uint16(src[pos:]))
+						mem.Disp = int64(int16(binary.LittleEndian.Uint16(src[pos:])))
 						pos += 2
 					}
 
@@ -611,7 +611,7 @@ Decode:
 					}
 					dispoff = pos
 					displen = 4
-					mem.Disp = int64(binary.LittleEndian.Uint32(src[pos:]))
+					mem.Disp = int64(int32(binary.LittleEndian.Uint32(src[pos:])))
 					pos += 4
 				}
